requirements/base.txt
Django==5
django-uuslug==2.0.0
psycopg2-binary==2.9.7
python-slugify==8.0.1
django-filter==23.3
djangorestframework==3.14.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
swagger-spec-validator==3.0.3